new changes, comments model updated

+++ RentHub/src/app/components/apartment/all-user-apartment/all-user-apartment.component.ts
@@ -48,15 +48,18 @@ export class AllUserApartmentComponent implements OnInit {
   }
 
   deletePost(id: number){
-    this.apartmentService.deleteData(id).subscribe(
-      (res: any) =>{
-        if(res.message == "Deleted"){
-          window.location.reload();
+    const op = window.confirm("Do you want to delete the apartment ?")
+    if(op) {
+      this.apartmentService.deleteData(id).subscribe(
+        (res: any) =>{
+          if(res.message == "Deleted"){
+            window.location.reload();
+          }
+        }, 
+        (err: any) => {
+          console.log(err)
         }
-      }, 
-      (err: any) => {
-        console.log(err)
-      }
-    )
+      )
+    }
   }
 }
